ESP32CamObjectDetection/remote/requirementVirtualenv.txt
absl-py==0.9.0
astor==0.8.1
click==7.1.1
cycler==0.10.0
Flask==1.1.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
itsdangerous==1.1.0
Jinja2==2.11.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
numpy==1.18.2
opencv-python==3.4.6.27
opt-einsum==3.2.0
Pillow==7.1.1
pkg-resources==0.0.0
protobuf==3.11.3
pyparsing==2.4.7
PyQt5==5.14.2
PyQt5-sip==12.7.2
pyserial==3.4
python-dateutil==2.8.1
PyYAML==5.3.1
scipy==1.4.1
six==1.14.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
Werkzeug==1.0.1
wrapt==1.12.1
